# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Aplikasi REST API",
    "description": "Dokumentasi yang dihasilkan dengan <b>swagger-autogen</b> module."
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/admin/": {
      "get": {
        "tags": [
          "Admin"
        ],
        "description": "Endpoint untuk mengambil semua Admin.",
        "responses": {
          "200": {
            "description": "Output API untuk objek Admin",
            "schema": {
              "$ref": "#/definitions/Admins"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Admin"
        ],
        "description": "Endpoint untuk membuat admin baru.",
        "parameters": [
          {
            "name": "data",
            "in": "body",
            "description": "Informasi Admin.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Admin"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Output API untuk objek Admin",
            "schema": {
              "$ref": "#/definitions/Admin"
            }
          }
        }
      }
    },
    "/admin/{id}": {
      "get": {
        "tags": [
          "Admin"
        ],
        "description": "Endpoint untuk mengambil Admin berdasarkan id.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Output API untuk objek Admin",
            "schema": {
              "$ref": "#/definitions/Admin"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Admin"
        ],
        "description": "Endpoint untuk index http put.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "data",
            "in": "body",
            "description": "Informasi Admin.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Admin"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Output API untuk objek Admin",
            "schema": {
              "$ref": "#/definitions/Admin"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Admin"
        ],
        "description": "Endpoint untuk menghapus Admin.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Output API untuk objek Admin",
            "schema": {
              "$ref": "#/definitions/Admin"
            }
          }
        }
      }
    },
...
    "/keamanan/": {
      "get": {
        "tags": [
          "Keamanan"
        ],
        "description": "Endpoint untuk mengambil semua Keamanan.",
        "responses": {
          "200": {
            "description": "Output API untuk objek Keamanan",
            "schema": {
              "$ref": "#/definitions/Keamanan"
            }
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "judul": {
                  "example": "any"
                },
                "lokasi": {
                  "example": "any"
                },
                "deskripsi": {
                  "example": "any"
                },
                "tanggal": {
                  "example": "any"
                },
                "notlp": {
                  "example": "any"
                },
                "fotoBase64": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/keamanan/{id}": {
      "get": {
        "tags": [
          "Keamanan"
        ],
        "description": "Endpoint untuk mengambil Keamanan berdasarkan id.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Output API untuk objek Keamanan",
            "schema": {
              "$ref": "#/definitions/Keamanan"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Keamanan"
        ],
        "description": "Endpoint untuk index http put.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "data",
            "in": "body",
            "description": "Informasi Keamanan.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Keamanan"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Output API untuk objek Keamanan",
            "schema": {
              "$ref": "#/definitions/Keamanan"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Keamanan"
        ],
        "description": "Endpoint untuk menghapus Keamanan.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Output API untuk objek Keamanan",
            "schema": {
              "$ref": "#/definitions/Keamanan"
            }
          }
        }
      }
    },
    "/kebakaran/": {
      "get": {
        "tags": [
          "Kebakaran"
        ],
        "description": "Endpoint untuk mengambil semua Kebakaran.",
        "responses": {
          "200": {
            "description": "Output API untuk objek Kebakaran",
            "schema": {
              "$ref": "#/definitions/Kebakaran"
            }
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "judul": {
                  "example": "any"
                },
                "lokasi": {
                  "example": "any"
                },
                "deskripsi": {
                  "example": "any"
                },
                "tanggal": {
                  "example": "any"
                },
                "notlp": {
                  "example": "any"
                },
                "fotoBase64": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/kebakaran/{id}": {
      "get": {
        "tags": [
          "Kebakaran"
        ],
        "description": "Endpoint untuk mengambil Kebakaran berdasarkan id.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Output API untuk objek Kebakaran",
            "schema": {
              "$ref": "#/definitions/Kebakaran"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Kebakaran"
        ],
        "description": "Endpoint untuk index http put.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "data",
            "in": "body",
            "description": "Informasi Kebakaran.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Kebakaran"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Output API untuk objek Kebakaran",
            "schema": {
              "$ref": "#/definitions/Kebakaran"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Kebakaran"
        ],
        "description": "Endpoint untuk menghapus Kebakaran.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Output API untuk objek Kebakaran",
            "schema": {
              "$ref": "#/definitions/Kebakaran"
            }
          }
        }
      }
    }
  },
  "definitions": {
...
    "Admin": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "nama": {
          "type": "string",
          "example": "budi"
        },
        "email": {
          "type": "string",
          "example": "budijati@gmail.com"
        },
        "password": {
          "type": "string",
          "example": "12345"
        }
      }
    },
    "Admins": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/Admin"
      }
    },
...
    "Kebakaran": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "Judul": {
          "type": "string",
          "example": "Banjir Bandang"
        },
        "lokasi": {
          "type": "string",
          "example": "Jl Menjangan IV"
        },
        "deskripsi": {
          "type": "string",
          "example": "Air Tidak Mengalir"
        },
        "tanggal": {
          "type": "string",
          "example": "Selasa 26 Novber 2021"
        },
        "notlp": {
          "type": "string",
          "example": "081387887675"
        },
        "foto": {
          "type": "string",
          "example": "cover.jpg"
        }
      }
    },
...
  }
}